Add numeric guards for stage1 BCE training

+++ losses/det_loss.py
@@ -3,4 +3,4 @@ import torch.nn.functional as F
 
 
 def detection_bce_loss(p_edit: torch.Tensor, target: torch.Tensor) -> torch.Tensor:
-    return F.binary_cross_entropy(p_edit, target.float())
+    return F.binary_cross_entropy(p_edit.float(), target.float())
