+++ losses/mask_loss.py
@@ -3,6 +3,8 @@ import torch.nn.functional as F
 
 
 def dice_loss(pred: torch.Tensor, target: torch.Tensor, eps: float = 1e-6) -> torch.Tensor:
+    pred = pred.float()
+    target = target.float()
     pred = pred.contiguous().view(pred.size(0), -1)
     target = target.contiguous().view(target.size(0), -1)
     inter = (pred * target).sum(dim=1)
@@ -12,5 +14,7 @@ def dice_loss(pred: torch.Tensor, target: torch.Tensor, eps: float = 1e-6) -> to
 
 
 def bce_dice_loss(pred: torch.Tensor, target: torch.Tensor) -> torch.Tensor:
-    bce = F.binary_cross_entropy(pred, target.float())
+    pred = pred.float()
+    target = target.float()
+    bce = F.binary_cross_entropy(pred, target)
     return bce + dice_loss(pred, target)
